% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Data.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A cleaned data frame.
}
\description{
Removes missing values and converts column names to lowercase.
}
