% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot Summary
Creates a bar plot from summarized data.}
\usage{
plot_summary(df, x, y)
}
\arguments{
\item{df}{A data frame.}

\item{x}{Column for x-axis.}

\item{y}{Column for y-axis.}
}
\value{
A ggplot2 bar chart.
}
\description{
Plot Summary
Creates a bar plot from summarized data.
}
