% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize_by_group.R
\name{summarize_by_group}
\alias{summarize_by_group}
\title{Summarize by Group}
\usage{
summarize_by_group(df, group_col, summary_col)
}
\arguments{
\item{df}{A data frame to summarize.}

\item{group_col}{string: the name of the column to group.}

\item{summary_col}{Column name to summarize.}
}
\value{
A summarized data frame.
}
\description{
Groups the data and calculates the mean of a selected column.
}
